feat: implement support for querying deeply nested remote table names (i.e. `SELECT * FROM one.two.three.four.five`) (#35)
* Add MultiPartTableReference

* Capture but don't replace MultiPartTableReferences in the LogicalPlan

* Implement rewrite logic for multi-part tables

* Fix comment

* merging

* Also merge

* Add function support

* Fix doctest running
